Tests/LibWeb/Text/expected/css/display-transitions-require-allow-discrete.txt

display without allow-discrete: none
transitions on display without allow-discrete: 0
content-visibility without allow-discrete: hidden
transitions on content-visibility without allow-discrete: 0
